fix wrong node fee display
Continuously update the node fee info. This should fix the wrong node fees that are sometimes displayed. I expect those wrong fees to be due to the connection starting and to be correct after a few seconds. Let us see with the users if that fixes the issue.
At the same time, it also avoids the potential exploit by node operators to suddenly restart their node with a higher fee.